pyproject.toml
[tool.poetry]
name = "src"
version = "0.64.1"
description = "GrooveShop Backend"
authors = [
    "Vasilis Totskas"
]

[tool.semantic_release]
version_variable = [
    "setup.py:__version__",
    "pyproject.toml:version"
]
branch = "master"
upload_to_pypi = false
upload_to_release = true
build_command = "pip install poetry && poetry build"
parser_angular_patch_types = "fix,feat,chore,perf,refactor,build,style,ci,test"
changelog_sections = "feature,fix,breaking,documentation,performance,refactor,build,style,ci,test,other"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0.0"
django = "4.2.3"
psycopg2-binary = "*"
djangorestframework = "3.14.0"
django-cors-headers = "4.1.0"
pillow = "10.0.0"
django-admin-thumbnails = "0.2.8"
django-mptt = "0.14.0"
django-js-asset = "2.1.0"
django-redis = "*"
Faker = "18.11.2"
django-environ = "0.10.0"
django-tinymce = "3.6.1"
gunicorn = "20.1.0"
whitenoise = "*"
django-filter = "23.2"
python-semantic-release = "7.34.6"
flake8 = "6.0.0"
coverage= "7.2.7"
pylint = "2.17.4"
pre-commit= "3.3.3"
django-allauth = "0.54.0"
django-allauth-2fa = "0.11.0"
django-celery-beat = "2.5.0"
django-celery-results = "2.5.1"
flower = "2.0.0"
djangorestframework-camel-case = "1.4.2"
djangoql = "*"
django-extensions = "*"
pytest = "*"
pytest-cov = "*"
sentry-sdk = "*"
shortuuid = "*"
unittest-xml-reporting = "*"
werkzeug = "*"

    [tool.poetry.dependencies.celery]
    version = ">=5.2.7,<6.0.0"
    extras = [ "redis" ]

[tool.poetry.dev-dependencies]
ipython = "*"
bpython = "*"
pudb = "*"
black = "23.3.0"

[tool.setuptools.packages.find]
where = ["src"]  # list of folders that contain the packages (["."] by default)
include = ["*"]  # package names should match these glob patterns (["*"] by default)
exclude = ["files*", "logs*"]  # exclude packages matching these glob patterns
namespaces = false  # to disable scanning PEP 420 namespaces (true by default)

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.coverage.paths]
source = ["*"]

[tool.coverage.run]
branch = true
omit = [
    "*/migrations/*",
    "*/management/*",
    "*/manage.py",
    "*/setup.py",
    "*/asgi.py",
    "*/wsgi.py",
]
source = ["."]
relative_files = true

[tool.coverage.report]
precision = 2
skip_covered = true
fail_under = 50
show_missing = true
